Fixing betty style

+++ 0-bubble_sort.c
@@ -6,12 +6,13 @@
  * @n2: int
  * Return: nothing
  */
+
 void cambio(int *n1, int *n2)
 {
 	int tmp = *n1;
+
 	*n1 = *n2;
 	*n2 = tmp;
-	
 }
 
 /**
@@ -20,21 +21,19 @@ void cambio(int *n1, int *n2)
  * @size: size_t
  * Return: nothing
  */
+
 void bubble_sort(int *array, size_t size)
 {
-	int i, j;
+	size_t i, j;
 
 	if (array == NULL)
 		printf("empy array");
-	
-	
+
 	for (i = 0; i < size - 1; i++)
 	{
-		
 		for (j = 0; j < size - i - 1; j++)
 		{
-
-			if(array[j] > array[j + 1])
+			if (array[j] > array[j + 1])
 			{
 				cambio(&array[j], &array[j + 1]);
 				print_array(array, size);
